+++ test/inference/prediction_tests.jl
@@ -16,9 +16,9 @@
     # They are essentially the same with the exception that the second model has multiple observations
     # Each case has two options: KeepLast() and KeepEach() which specifies how many predictions we should keep for the result
     for a in (1, 2, 3), b in (2, 3, 4), iterations in (10, 20, 30)
-        expected_prediction = call_message_update_rule(
+        expected_prediction = getresult(call_message_update_rule(
             Bernoulli, :out; q = (p = Beta(a, b),)
-        )
+        ))
 
         @testset "beta_bernoulli_single" for option in (KeepLast(), KeepEach())
             result_missing_data_explicitly_specified_1 = infer(
@@ -67,9 +67,9 @@
 
         @testset "beta_bernoulli_multiple" begin
             for n in (10, 20, 30), option in (KeepLast(), KeepEach())
-                expected_prediction = call_message_update_rule(
+                expected_prediction = getresult(call_message_update_rule(
                     Bernoulli, :out; q = (p = Beta(a, b),)
-                )
+                ))
 
                 result_missing_data_explicitly_specified_1 = infer(
                     model = beta_bernoulli_multiple(a = a, b = b, n = n),